requirements.txt
numpy >= 1.22.4
Pillow >= 9.1.1
scipy >= 1.8.1